Introduce `MenuBarItemSourceCache`
This cache essentially allows us to determine the true `ownerPID` for menu bar items, since the process that owns all menu bar items in macOS Tahoe is the Control Center. We're using `sourcePID` to refer to this separately, since `ownerPID` could still be useful in some way.

NOTE: A bunch of other changes got mixed into this commit, and I didn't realize until it was too deep in the commit list to fix. Here are some of the more notable things that were added, aside from `MenuBarItemSourceCache`:

- Refactor `Bridging` APIs
- Update `MenuBarItem` implementation
- Replace `MenuBarItemInfo` and `MenuBarItemLegacyInfo` with `MenuBarItemTag` (essentially the original `MenuBarItemInfo`, but using `sourcePID` instead of `ownerPID`)
- Improve `MenuBarItem` image caching
- Change some documentation wording
- Misc UI changes